refactor: de-duplicate ArgumentExceptionHelper into a single net4x polyfill
SonarCloud flagged the two near-identical ArgumentExceptionHelper copies
(Async + Extensions, ~89 duplicated lines each). Replace them with a single
shared polyfill under Polyfills/ that is compiled only on net462-net481, and
on net8.0+ alias the `ArgumentExceptionHelper` identifier directly to
`System.ArgumentNullException` so the 250 `ThrowIfNull` call sites bind to the
BCL method with no shipped wrapper.

- Drop the dead `ThrowIfNullOrEmpty` (0 uses) and inline the 3 `Check<T>`
  uses (in CombineLatestEnumerable) as `?? throw`.
- Remove the now-redundant `using ...Internal;` imports the alias orphaned.
- Remove 3 no-op `ThrowIfNull(ImmutableArray<>)` guards CA2264 surfaced once
  the alias bound them to the real BCL method (a struct is never null).

Async 1215 and Extensions 687 tests pass; both build clean across all eight
TFMs with -warnaserror.
