+++ src/ReactiveUI.Primitives.Async/Observables/Throw.cs
@@ -5,7 +5,6 @@
 using System.Diagnostics.CodeAnalysis;
 
 using ReactiveUI.Primitives.Async.Disposables;
-using ReactiveUI.Primitives.Internal;
 
 namespace ReactiveUI.Primitives.Async;
