+++ src/ReactiveUI.Primitives.Async/Operators/Zip.cs
@@ -4,7 +4,6 @@
 
 using ReactiveUI.Primitives.Async.Disposables;
 using ReactiveUI.Primitives.Async.Internals;
-using ReactiveUI.Primitives.Internal;
 
 namespace ReactiveUI.Primitives.Async;
