+++ src/ReactiveUI.Primitives.Async/Signals/Base/BaseStatelessReplayLatestSignalAsync.cs
@@ -5,7 +5,6 @@
 using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using ReactiveUI.Primitives.Async.Internals;
-using ReactiveUI.Primitives.Internal;
 
 namespace ReactiveUI.Primitives.Async.Signals;
