+++ src/ReactiveUI.Primitives.Extensions/ObservableSubscriptionExtensions.cs
@@ -4,7 +4,6 @@
 
 using ReactiveUI.Primitives.Concurrency;
 using ReactiveUI.Primitives.Disposables;
-using ReactiveUI.Primitives.Extensions.Internal;
 
 namespace ReactiveUI.Primitives.Extensions;
