+++ src/ReactiveUI.Primitives.Extensions/ObserverExtensions.cs
@@ -2,8 +2,6 @@
 // ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-using ReactiveUI.Primitives.Extensions.Internal;
-
 namespace ReactiveUI.Primitives.Extensions;
 
 /// <summary>Extension methods for <see cref="IObserver{T}"/>.</summary>
